fix (nav): mobile create btn, fix default nav data
