+++ src/features/Posts/pages/FeedPage/styles.module.scss
@@ -42,11 +42,7 @@
     min-width: 0;
   }
 
-  grid-template-areas:
-    "title"
-    "categories"
-    "sort-by"
-    "content";
+  grid-template-areas: "content";
 
   :global {
     #title {
