+++ src/features/Projects/pages/ExplorePage/ExplorePage.tsx
@@ -6,7 +6,7 @@ import SkipLink from "src/Components/SkipLink/SkipLink";
 
 export default function ExplorePage() {
   return (
-    <div className="bg-white p-16 rounded-24">
+    <div>
       <SkipLink id="explore-projects-content">Skip To Content</SkipLink>
       <OgTags title="#BuildOnBitcoin" />
       <main id="explore-projects-content">
